refactor phedexinjector, fixes #6451
Move all PhEDExInjector functionality into a single thread. Also remove all transactions and cleanup/simplify code a bit.
