+++ src/python/WMComponent/DBS3Buffer/MySQL/DBSBufferFiles/SetPhEDExStatus.py
@@ -10,8 +10,9 @@ from WMCore.Database.DBFormatter import DBFormatter
 class SetPhEDExStatus(DBFormatter):
     sql = "UPDATE dbsbuffer_file SET in_phedex = :status WHERE lfn = :lfn"
 
-    def execute(self, lfns, status, conn = None, transaction = None):
-        if type(lfns) != list:
+    def execute(self, lfns, status, conn = None, transaction = False):
+
+        if not isinstance(lfns, list):
             lfns = [lfns]
 
         if len(lfns) < 1:
@@ -21,6 +22,7 @@ class SetPhEDExStatus(DBFormatter):
         for lfn in lfns:
             bindVars.append({"lfn": lfn, "status": status})
 
-        result = self.dbi.processData(self.sql, bindVars, conn = conn,
-                                      transaction = transaction)
+        self.dbi.processData(self.sql, bindVars, conn = conn,
+                             transaction = transaction)
+
         return
